requirements.txt
faiss-cpu==1.13.2
gradio==6.12.0
groq==0.37.1
huggingface_hub==1.11.0
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.2.31
langchain-groq==1.1.2
langchain-huggingface==1.2.2
langchain-text-splitters==1.1.2
pypdf==6.10.2
python-dotenv==1.2.2
sentence-transformers==5.4.1